requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==5.0.1
python-decouple==3.8
sqlparse==0.5.1
